rename file to useUIMessages
